Migrate bottom bar to header
- Restyle header
- Create new login-register page
- Enable guest user to see the item list
