+++ frontend/simple-mercari-web/src/App.tsx
@@ -1,13 +1,14 @@
-import React, { useState } from "react"
-import { Routes, Route, BrowserRouter } from "react-router-dom"
-import { Home } from "./components/Home"
-import { ItemDetail } from "./components/ItemDetail"
-import { UserProfile } from "./components/UserProfile"
-import { Listing } from "./components/Listing"
-import "./App.css"
-import { Header } from "./components/Header/Header"
-import { ToastContainer } from "react-toastify"
-import "react-toastify/dist/ReactToastify.css"
+import React, { useState } from "react";
+import { Routes, Route, BrowserRouter } from "react-router-dom";
+import { Home } from "./components/Home";
+import { ItemDetail } from "./components/ItemDetail";
+import { UserProfile } from "./components/UserProfile";
+import { Listing } from "./components/Listing";
+import "./App.css";
+import { Header } from "./components/Header/Header";
+import { ToastContainer } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
+import { SignUpAndSignIn } from "./components/SignInSignUp";
 
 export const App: React.FC = () => {
   const [searchValue, setSearchValue] = useState<string>("")
@@ -26,6 +27,7 @@ export const App: React.FC = () => {
             <Route path="/item/:id" element={<ItemDetail />} />
             <Route path="/user/:id" element={<UserProfile />} />
             <Route path="/sell" element={<Listing />} />
+            <Route path="/login" element={<SignUpAndSignIn />} />
           </Routes>
         </div>
       </BrowserRouter>
