+++ frontend/simple-mercari-web/src/components/Header/Header.tsx
@@ -1,18 +1,22 @@
-import { useCookies } from "react-cookie"
-import "./Header.css"
-import { Navbar, Container, Nav } from "react-bootstrap"
-import { FaSearch } from "react-icons/fa"
-import { AiOutlineClose } from "react-icons/ai"
-import { useState } from "react"
-import { SearchBar } from "../SearchBar"
+import { useCookies } from "react-cookie";
+import "./Header.css";
+import { Navbar, Container, Nav, Form, FormControl, Button } from 'react-bootstrap';
+import { FaSearch } from 'react-icons/fa';
+import { AiOutlineClose } from 'react-icons/ai';
+import { useState } from "react";
+import { SearchBar } from "../SearchBar/SearchBar";
+import { useNavigate } from "react-router-dom";
+
 
 interface SearchBarProps {
   onSearch: (keyword: string) => void
 }
 
 export const Header = ({ onSearch }: SearchBarProps) => {
-  const [cookies, _, removeCookie] = useCookies(["userID", "token"])
-  const [showNavbar, setShowNavbar] = useState(false)
+  const [cookies, _, removeCookie] = useCookies(["userID", "token"]);
+  const [showNavbar, setShowNavbar] = useState(false);
+  const navigate = useNavigate();
+
 
   const onLogout = (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
     event.preventDefault()
@@ -28,29 +32,23 @@ export const Header = ({ onSearch }: SearchBarProps) => {
       <Navbar bg="#222427" variant="dark" expand="lg">
         <Container fluid>
           <Nav>
-            {showNavbar ? null : <Navbar.Brand>Simple Mercari</Navbar.Brand>}
-            <Navbar.Toggle
-              aria-controls="responsive-navbar-nav"
-              onClick={handleNavbarToggle}
-            >
-              {" "}
-              {showNavbar ? <AiOutlineClose /> : <FaSearch />}
-            </Navbar.Toggle>
-            <Navbar.Collapse
-              id="responsive-navbar-nav"
-              className={showNavbar ? "show" : ""}
-            >
+            {showNavbar ? null : <Navbar.Brand className="hand-cursor" onClick={() => navigate("/")}>Simple Mercari</Navbar.Brand>}
+            <Navbar.Toggle aria-controls="responsive-navbar-nav" onClick={handleNavbarToggle}> {showNavbar ? <AiOutlineClose /> : <FaSearch />}</Navbar.Toggle>
+            <Navbar.Collapse id="responsive-navbar-nav" className={showNavbar ? 'show' : ''}>
               <Nav className="me-auto">
                 <SearchBar onSearch={onSearch} />
               </Nav>
             </Navbar.Collapse>
           </Nav>
-          {showNavbar ? null : (
-            <Nav className="md:ml-auto">
-              <Nav.Link href="#account">Account</Nav.Link>
-              <Nav.Link onClick={onLogout}>Logout</Nav.Link>
+          {
+            showNavbar ? null : <Nav className="md:ml-auto">
+
+              {cookies.token ||
+                cookies.userID ? <Nav> <Nav.Link onClick={() => navigate(`/user/${cookies.userID}`)}>Account</Nav.Link>
+                <Nav.Link onClick={onLogout}>Logout</Nav.Link> <button id="MerButton" onClick={() => navigate("/sell")}>Sell</button> </Nav> : <Nav><Nav.Link onClick={() => navigate("/login")}>Account</Nav.Link><button id="MerButton" onClick={() => navigate("/login")}>Sell</button> </Nav>}
+
             </Nav>
-          )}
+          }
         </Container>
       </Navbar>
     </>
