+++ frontend/simple-mercari-web/src/components/MerComponent/MerComponent.tsx
@@ -10,18 +10,9 @@ interface Prop {
 
 export const MerComponent: React.FC<Prop> = (props) => {
   const [cookies] = useCookies(["token", "userID"]);
-
-  if (
-    !cookies.token ||
-    !cookies.userID ||
-    (props.condition && props.condition() === false)
-  ) {
-    return <NotFound />;
-  }
   return (
     <>
       {props.children}
-      <Footer />
     </>
   );
 };
