+++ frontend/simple-mercari-web/src/components/SignInSignUp/index.tsx
@@ -1,0 +1,1 @@
+export * from "./SignInSignUp";
